More renaming, sia --> sentient
