+++ algorithms/sentient/kernel.go
@@ -1,4 +1,4 @@
-package sia
+package sentient
 
 const kernelSource = `
