+++ algorithms/sentient/miner.go
@@ -1,4 +1,4 @@
-package sia
+package sentient
 
 import (
 	"log"
