+++ algorithms/sentient/miner_test.go
@@ -1,4 +1,4 @@
-package sia
+package sentient
 
 import (
 	"bytes"
